Pass absolute paths to Inkscape
Inkscape may be called through a wrapper script which changes the
current working directory. If so, Inkscape operations will fail because
it won't find the specified files.

On the Mac, this is definitely required if using the official Inkscape
distribution. I'm not sure about other platforms but none should stop
working with an absolute path. It means the messges output by Inkscape
during image generation have longer and uglier paths, though - maybe in
future we should trap those and prettify them.