Move functions out of pcbmode.utils.svg
To remove duplicate code in pcbmode.utils.svg, it needs to use the
methods in pcbmode.utils.svgpath - that would set up a dependency cycle
so this move needs to happen first. This enables .svg to use .svgpath
instead of the other way around, and the functions can be converted to
proper methods later.